microservices/images/python/requirements.txt
python-resize-image
Redis
pika
requests